feat: Add get match data call with unit test

+++ app/src/main/java/com/example/opgg/data/summoner/SummonerData.kt
@@ -217,7 +217,7 @@ class MatchData {
         val assist: Int,
         val kdaString: String,
         val cs: Int,
-        val csPerMin: Int,
+        val csPerMin: Float,
         val contributionForKillRate: String,
         val goldEarned: Int,
         val totalDamageDealtToChampions: Int,
